+++ CHANGELOG.md
@@ -39,6 +39,7 @@
 * Change the format of the input file, to make it more robust and adaptative (#332)
 * Input Editor window can get behind main window (#334, #49)
 * Home/End keys in the input editor move to top/bottom row (#335)
+* Use periodic UI update for performance
 
 ### Fixed